requirements.txt
cachetools==3.0.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
mccabe==0.6.1
mysql-connector-python==8.0.13
oauth2client==4.1.3
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
rsa==4.0
six==1.12.0
uritemplate==3.0.0
